Push docker image to DockerHub on merge into master
The build process will take around 20 and 10 minutes when building a 
native image on the native arch for gcc and sail, respectively, if 
internet connection is fast

The multi-arch image build will take a long time (4 hours on a 8-core 
machine with fast internet) as we will use QEMU emulation for cross 
compilation. At the time of writing, we can only run multi-arch build 
successfully on x86 machines.